Add files via upload
To begin, we'll build the chat_server.py. What will the functions of our server be?

First, the server needs to accept new connections from clients. From here, we need to come up with some way to identify our unique users. We could display users by IP address, but most people tend to rather come up with some sort of username, so our server will first allow clients to connect and choose a username. Beyond this, the server will collect incoming messages and then distribute them to the rest of the connected clients.

Server Code
In our architecture, the whole job of the server is to do two essential operations:

Listening for upcoming client connections, if a new client is connected, we add it to our collection of client sockets.
Start a new thread for each client connected that keep listening for upcoming messages sent from the client and broadcast it to all other clients.


Client Code
The client does three basic operations:

Connects to the server.
Keep listening for messages coming from the server (must be a client sent a message to the server and the server broadcasted it) and print it to the console.
Waiting for user to input messages to send to the server.